register error part

+++ src/features/counter/counterSlice.js
@@ -9,13 +9,27 @@ export const counterSlice = createSlice({
   initialState,
   reducers: {
     addToCart: (state, action) => {
-      state.product.push(action.payload);
-      localStorage.setItem("product", JSON.stringify(state.product));
+      const existingProduct = state.product.find(
+        (item) => item.id === action.payload.id
+      );
+      if (existingProduct) {
+        existingProduct.quantity =
+           action.payload.quantity;
+        existingProduct.allPrice =
+          existingProduct.quantity * existingProduct.price;
+        localStorage.setItem("product", JSON.stringify(state.product));
+      } else {
+        state.product.push(action.payload);
+        localStorage.setItem("product", JSON.stringify(state.product));
+      }
     },
     removeFromCart: (state, action) => {
-      state.product = state.product.filter((item)=> item.id !== action.payload);
-    }
-  }
+      state.product = state.product.filter(
+        (item) => item.id !== action.payload
+      );
+      localStorage.setItem("product", JSON.stringify(state.product));
+    },
+  },
 });
 
 // Action creators are generated for each case reducer function
